+++ modules/router/src/test/forwarding.spec.ts
@@ -9,7 +9,6 @@ import {
   mkPublicIdentifier,
   RestServerNodeService,
 } from "@connext/vector-utils";
-import { VectorChainReader } from "@connext/vector-contracts";
 import { AddressZero } from "@ethersproject/constants";
 import pino from "pino";
 import Sinon from "sinon";
@@ -23,7 +22,6 @@ import { mockProvider } from "./utils/mocks";
 const hydratedProviders = { 1337: mockProvider };
 
 const logger = pino({ level: config.logLevel });
-const chainReader = new VectorChainReader(hydratedProviders, logger);
 
 describe("Forwarding", () => {
   describe("transferCreation", () => {
@@ -76,7 +74,7 @@ describe("Forwarding", () => {
         bob: mkAddress("0xb1"),
         channelAddress: data.channelAddress,
         balances: [data.channelBalance],
-      });
+      }).channel;
       const receiverChannel = createTestChannelState("deposit", {
         alice: mkAddress("0xa"),
         bob: mkAddress("0xb2"),
@@ -87,7 +85,7 @@ describe("Forwarding", () => {
             to: [mkAddress("0xb"), mkAddress("0xc")],
           },
         ],
-      });
+      }).channel;
       node.getStateChannel.resolves(Result.ok(senderChannel));
       node.getStateChannelByParticipants.resolves(Result.ok(receiverChannel));
       node.conditionalTransfer.resolves(Result.ok({} as any));