+++ modules/documentation/docs/changelog.md
@@ -2,7 +2,17 @@
 
 ## Next Release
 
+## 0.0.32
+
+- add utility sign method
+
+## 0.0.31
+
 - fix browser node store issue
+- fix `TFullTransferState` schema
+- store `getChannelStateByParticipants` accepts identifiers instead of addresses
+- make sure all rpc queries properly return `Result` types in error cases
+- add `crossChainTransfer` to iframe
 
 ## 0.0.28